# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gudang documentation",
        "description": "The Documentation of API Gudang",
        "termsOfService": "http://swagger.io/terms/",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/categories/": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Return all category",
                "description": "Return all category with pagination",
                "operationId": "getCategories",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Return detail category",
                "description": "Return detail category by Id",
                "operationId": "getCategoriesById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Categories to use",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Category",
            "description": "Manage category data"
        }
    ]
}